RFS-68: Crash when Connection gets GCed.
Node disposes the object handle first and then deletes the object what puts
the object in not-fully-valid state during destructor call. The handle()
call returns empty handle and the hell breaks loose.

Solved by getting log function handle in constructors once and calling
it on themselves instead of the Connection/Function objects as it doesn't
really need an object (this value).